fix debug trace to compute egp percentage based on tx execution values
