+++ jsonrpc/types/interfaces.go
@@ -24,7 +24,6 @@ type PoolInterface interface {
 	GetTransactionByHash(ctx context.Context, hash common.Hash) (*pool.Transaction, error)
 	GetTransactionByL2Hash(ctx context.Context, hash common.Hash) (*pool.Transaction, error)
 	CalculateEffectiveGasPrice(rawTx []byte, txGasPrice *big.Int, txGasUsed uint64, l1GasPrice uint64, l2GasPrice uint64) (*big.Int, error)
-	CalculateEffectiveGasPricePercentage(gasPrice *big.Int, effectiveGasPrice *big.Int) (uint8, error)
 	EffectiveGasPriceEnabled() bool
 }
